backend/requirements.txt
fastapi
uvicorn[standard]
grpcio-tools